Fix Cora loading test on Windows (#1705)
NumPy translates `int` to `np.int_` equivalent to the C `long` type. This type
has different behaviour and size on different platforms, in particular Windows
(LLP64, becomes `np.int32`) vs. Linux/macOS (LP64, becomes `np.int64`):
https://en.wikipedia.org/wiki/64-bit_computing#64-bit_data_models

For the test here, the underlying data always ends up as `int64`, so we can just
test that.

See: #1698

+++ tests/datasets/test_datasets.py
@@ -193,7 +193,7 @@ def test_cora_load(is_directed, largest_cc_only, subject_as_feature) -> None:
     else:
         feature_size = base_feature_size
 
-    assert g.nodes().dtype == int
+    assert g.nodes().dtype == np.int64
     assert g.is_directed() == is_directed
 
     assert g.number_of_nodes() == expected_nodes
